+++ libpkg/private/utils.h
@@ -88,8 +88,8 @@ void sbuf_free(struct sbuf *);
 ssize_t sbuf_size(struct sbuf *);
 
 int mkdirs(const char *path);
-int file_to_buffer(const char *, char **, size_t *);
-int file_to_bufferat(int, const char *, char **, size_t *);
+int file_to_buffer(const char *, char **, off_t *);
+int file_to_bufferat(int, const char *, char **, off_t *);
 int format_exec_cmd(char **, const char *, const char *, const char *, char *,
     int argc, char **argv);
 int is_dir(const char *);
